requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
aiosqlite==0.19.0
aiohttp==3.9.1
flask>=2.3.0  # For health checks